Fix crash when building with g++ 4.3 or newer
This adds a true copy constructor to StdBuf and StdStrBuf.
Unfortunately, this constructor is sometimes prefered to the
move constructor when it wasn't with older compilers, always
referencing the original instead of copying makes the engine
crash, and omitting the copy constructor just creates an
implicit one which also causes crashes.

+++ engine/src/C4NetIO.cpp
@@ -2145,7 +2145,7 @@ void C4NetIOUDP::OnPacket(const C4NetIOPacket &Packet, C4NetIO *pNetIO)
 		// ping? answer without creating a connection
 		if((Packet.getStatus() & 0x7F) == IPID_Ping)
 		{
-			PacketHdr PingPacket = { IPID_Ping | (Packet.getStatus() & 0x80), 0 };
+			PacketHdr PingPacket = { int8_t(IPID_Ping | (Packet.getStatus() & 0x80)), 0 };
 			SendDirect(C4NetIOPacket(&PingPacket, sizeof(PingPacket), false, Packet.getAddr()));
 			return;
 		}
